[FLINK-8959][tests] Port AccumulatorLiveITCase to flip6
